Article feedback
I started writing feedback to you on Discord but then it started getting out of hand so I thought maybe this would be better. My first time using the GitHub UI to create a PR like this! Hope it works out OK 😅. First of all I just want to say that the article is incredible - a masterpiece, really. I imagine that it will be soon be the most consistently visited page on the blog - I love the iterative approach to building things out, and I think literally going from 3 lines of code to 400 works to really draw the reader in. Bravo!!

I had one piece of miscellaneous feedback that didn't fit into the category of "proposed" changes but had to be filed somewhere:

- "For now, the partition name is hardcoded 😭" - no comment other than to say that the emoji usage here was too good to not say something